# Conflicts:
#	frontend/src/main/scala/org/seekloud/thor/front/pages/ThorRender.scala
#	shared/src/main/scala/org/seekloud/thor/shared/ptcl/model/Constants.scala

+++ backend/src/main/scala/org/seekloud/thor/core/RoomActor.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2018 seekloud (https://github.com/seekloud)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.seekloud.thor.core
 
 import java.util.concurrent.atomic.AtomicLong
@@ -73,9 +89,9 @@ object RoomActor {
             implicit val sendBuffer: MiddleBufferInJvm = new MiddleBufferInJvm(81920)
             val thorSchema = ThorSchemaServerImpl(AppSettings.thorGameConfig, ctx.self, timer, log, dispatch(subscribersMap, watchingMap), dispatchTo(subscribersMap, watchingMap))
 
-            for (cnt <- 0 until thorSchema.config.getRobotNumber) {
+            for (cnt <- 0 until thorSchema.config.getRobotNumber){
               val tmpId = getTmpId(s"robot$cnt", thorSchema.config.getRobotNames(cnt), thorSchema)
-              ctx.self ! CreateRobot(s"robot$cnt", tmpId.toByte, thorSchema.config.getRobotNames(cnt), thorSchema.config.getRobotLevel)
+              ctx.self ! CreateRobot(s"robot$cnt", tmpId, thorSchema.config.getRobotNames(cnt), thorSchema.config.getRobotLevel)
             }
 
             if (AppSettings.gameRecordIsWork) {
