+++ shared/src/main/scala/org/seekloud/thor/shared/ptcl/component/Adventurer.scala
@@ -1,3 +1,19 @@
+/*
+ * Copyright 2018 seekloud (https://github.com/seekloud)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package org.seekloud.thor.shared.ptcl.component
 
 import org.seekloud.thor.shared.ptcl.config.ThorGameConfig
@@ -20,7 +36,7 @@ case class AdventurerState(
   killNum: Short,
   isMove: Byte,
   isUpdateLevel: Byte,
-  levelUpExecute: Short,
+  levelUpExecute: Byte,
   isIntersect: Byte
 )
 
@@ -164,7 +180,7 @@ trait Adventurer extends CircleObjectOfGame {
       killNum.toShort,
       if(isMove) 1 else 0,
       if(isUpdateLevel) 1 else 0,
-      levelUpExecute.toShort,
+      levelUpExecute.toByte,
       isIntersect)
   }
